refactor: format 관련 util 정리
