+++ src/components/DashboardPage/DashboardTask.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 import useGetTasksToday from '@/apis/dashboard/tasksToday/query';
 import BtnTask from '@/components/common/BtnTask/BtnTask';
 import { TaskType } from '@/types/tasks/taskType';
-import formatDatetoLocalDate from '@/utils/formatDatetoLocalDate';
+import { formatDatetoLocalDate } from '@/utils/formatDateTime';
 
 interface DashboardTaskProps {
 	taskStatus: string;
